requirements.txt
streamlit==1.50.0
tensorflow==2.15.0
numpy==1.24.4
pillow==11.2.1
pandas==2.3.1
